path2pgm: Fix accidental cast of coords to ints.
The cast was simplifying zrusin-poly2.path in particular
radically so were were getting much better render times.
This change drops the rendering of that path from 1420
to 750 fps.  Other paths aren't as badly affected since
their mean lineto length is more than a pixel.

+++ path2pgm.c
@@ -288,7 +288,7 @@ program_emit_int(struct program *p, int i)
 }
 
 static void
-program_emit_double(struct program *p, int x)
+program_emit_double(struct program *p, double x)
 {
         union mem c;
         c.double_ = x;
